Update to fix token signin flash on browser with expired access tokens
